Dockerize - Nginx, php fpm & mysql db  (#23)
* Dockerized app with php fpm nginx and mysql db

* Finished dockerize

* docker image php fpm

+++ database/migrations/2026_05_20_182343_create_post_comments_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -13,8 +12,8 @@ return new class extends Migration
     {
         Schema::create('post_comments', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('post_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('post_id');
             $table->text('comment');
             $table->enum('status', ['draft', 'published', 'taken_down'])->default('draft');
             $table->timestamps();
